replaced List/Select conponent by  Accordian to assign Transportation Operator
