+++ src/components/bco_bookings/assign_transportation_operator.tsx
@@ -3,107 +3,163 @@ import { DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTit
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "../ui/tooltip";
 import { Button } from "../ui/button";
 import { Label, Input } from "@aws-amplify/ui-react";
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectGroup, SelectLabel } from "@radix-ui/react-select";
-import { Loader2 } from "lucide-react";
 import { User } from "../users/columns";
+import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from "../ui/accordion";
+import { useState } from "react";
+import { BCODataTableMeta } from "./data-table";
 
-export const assignTransportationOperator = (isDialogOpen: any, row: any, handleOpen: any, setIsDialogOpen: any,
-                         tempEmail: any, tempName: any, isLoading: any, handleOperatorSelect: any,
-                         data: User[], getSelectItem: any, handleSubmit: any) =>{
 
-  
-return (
-    <>
-    <Dialog 
-                  open={isDialogOpen} 
-                  onOpenChange={(open) => {
-                    setIsDialogOpen(open)
+export const assignTransportationOperator = (table: any, row: any) => {
+
+        const [data, setData] = useState<User[]>([])
+        const [tempName, setTempName] = useState("")
+        const [tempEmail, setTempEmail] = useState("")
+        const [isDialogOpen, setIsDialogOpen] = useState(false)
+       
+        const handleOpen = async () => {
+          setIsDialogOpen(true)
+          const result = await (table.options.meta as BCODataTableMeta)?.fetchTransportationOperators();
+          setData(result)
+        }
+
+        const handleOperatorSelect = (value: string) => {
+          setTempEmail(value);
+          const selectedUser = data.find(
+            (user) => `${user.email}` === value
+          );
+          if (selectedUser) {
+            setTempName( `${selectedUser.given_name} ${selectedUser.family_name}`);
+          }
+        };
+
+        function handleSubmit() {
+          // Use the parent's updateCargo method:
+          (table.options.meta as BCODataTableMeta)?.assignTransOp(row.original.containerID, tempName, tempEmail, "Pending Transportation Operator Approval")
+          setIsDialogOpen(false)
+        }
+
+        const getSelectItem = (user: User) => {
+            const fullName = `${user.given_name} ${user.family_name}`;
+            return (
+                <p>
+                    <Label  onClick={()=>handleOperatorSelect(user.email)}>{user["custom:organization"] ? user["custom:organization"] : fullName}</Label>
                     
-                  }}
-                >
-                  <DialogTrigger asChild>
+                </p>
+
+            );
+        };
+
+    return (
+        <>
+            
+            <Dialog
+                open={isDialogOpen}
+                onOpenChange={(open) => {
+                    setIsDialogOpen(open)
+
+                }}
+            >
+                <DialogTrigger asChild>
                     <TooltipProvider>
-                      <Tooltip delayDuration={300}>
-                        <TooltipTrigger>
-                         <Button onClick={handleOpen} variant="outline" disabled={row.original.containerStatus === "On-Ship"} className="bg-blue-600 text-white hover:bg-blue-700">Assign</Button>
-                        </TooltipTrigger>
-                        {row.original.containerStatus === "On-Ship" && (
-                          <TooltipContent>
-                            <p>Container still on ship. Cannot assign operator yet.</p>
-                          </TooltipContent>
-                        )}
-                      </Tooltip>
+                        <Tooltip delayDuration={300}>
+                            <TooltipTrigger>
+                                <Button onClick={handleOpen} variant="outline" disabled={row.original.containerStatus === "On-Ship"} className="bg-blue-600 text-white hover:bg-blue-700">Assign</Button>
+                            </TooltipTrigger>
+                            {row.original.containerStatus === "On-Ship" && (
+                                <TooltipContent>
+                                    <p>Container still on ship. Cannot assign operator yet.</p>
+                                </TooltipContent>
+                            )}
+                        </Tooltip>
                     </TooltipProvider>
-                  </DialogTrigger>
-                  <DialogContent>
+                </DialogTrigger>
+                <DialogContent>
                     <DialogHeader>
-                      <DialogTitle>Assign Transportation Operator</DialogTitle>
-                      <DialogDescription>
-                      Enter a Transportation Operator name and email to assign this container.
-                      </DialogDescription>
+                        <DialogTitle>Assign Transportation Operator</DialogTitle>
+                        <DialogDescription>
+                            Enter a Transportation Operator name and email to assign this container.
+                        </DialogDescription>
                     </DialogHeader>
                     <div className="space-y-2 py-2">
-                      <div>
-                        <Label>Transportation Operator</Label>
-                        <Select value={tempEmail} onValueChange={handleOperatorSelect} disabled={isLoading}>
-                          <SelectTrigger className="w-full">
-                            <div className="flex items-center gap-2">
-                              {isLoading && <Loader2 className="h-4 w-4 animate-spin" />}
-                              <SelectValue placeholder="Select operator" >
-                              {tempName}  
-                              </SelectValue>
-                            </div>
-                          </SelectTrigger>
-                          <SelectContent>
-                            <SelectGroup>
-                              <SelectLabel>Trucking Operators</SelectLabel>
-                              {data.map(user => {
-                                if ('Trucking Operator' === `${user["custom:role"]}`) {
-                                      return getSelectItem(user);
-                                }
-                              })}
-                            </SelectGroup>
-                            <SelectGroup>
-                              <SelectLabel>Rail Operators</SelectLabel>
-                              {data.map(user => {
-                                if ('Rail Operator' === `${user["custom:role"]}`) {
-                                      return getSelectItem(user);
-                                }
-                              })}
-                            </SelectGroup>
-                            <SelectGroup>
-                              <SelectLabel>Third Party Logistic Providers</SelectLabel>
-                              {data.map(user => {
-                                
-                                if ('Third Party Logistics Provider' === `${user["custom:role"]}`) {
-                                      return getSelectItem(user);
-                                }
-                              })}
-                            </SelectGroup>
-                          </SelectContent>
-                        </Select>
-                      </div>
-                      <div>
-                        <Label>Transportation Operator Email</Label>
-                        <Input
-                          value={tempEmail}
-                          disabled={true}
-                        />
-                      </div>
+                        <div>
+                            <Label>Transportation Operator</Label>
+                            <Accordion
+                                type="single"
+                                collapsible
+                                className="w-full"
+                            >
+                                <AccordionItem value="item-1">
+
+
+                                    <AccordionTrigger>
+                                        Trucking Operators
+                                    </AccordionTrigger>
+                                    <AccordionContent className="flex flex-col gap-4 text-balance">
+                                        <p>
+                                            {data.map(user => {
+                                                if ('Trucking Operator' === `${user["custom:role"]}`) {
+                                                    return getSelectItem(user);
+                                                }
+                                            })}
+                                        </p>
+                                    </AccordionContent>
+                                </AccordionItem>
+                                <AccordionItem value="item-2">
+
+                                    <AccordionTrigger>
+                                        Rail Operators
+                                    </AccordionTrigger>
+                                    <AccordionContent className="flex flex-col gap-4 text-balance">
+
+                                        <p>
+                                            {data.map(user => {
+                                                if ('Rail Operator' === `${user["custom:role"]}`) {
+                                                    return getSelectItem(user);
+                                                }
+                                            })}
+                                        </p>
+                                    </AccordionContent>
+                                </AccordionItem>
+                                <AccordionItem value="item-3">
+
+                                    <AccordionTrigger>
+                                        Third Party Logistic Providers
+                                    </AccordionTrigger>
+                                    <AccordionContent className="flex flex-col gap-4 text-balance">
+                                        <p>
+                                            {data.map(user => {
+
+                                                if ('Third Party Logistics Provider' === `${user["custom:role"]}`) {
+                                                    return getSelectItem(user);
+                                                }
+                                            })}
+                                        </p>
+                                    </AccordionContent>
+                                </AccordionItem>
+
+                            </Accordion>
+                        </div>
+                        <div>
+                            <Label>Transportation Operator Email</Label>
+                            <Input
+                                value={tempEmail}
+                                disabled={true}
+                            />
+                        </div>
                     </div>
                     <DialogFooter>
-                      <Button variant="outline" onClick={() => setIsDialogOpen(false)}>
-                        Cancel
-                      </Button>
-                      <Button onClick={handleSubmit} disabled={!tempName.trim() || !tempEmail.trim()}>
-                        Submit
+                        <Button variant="outline" onClick={() => setIsDialogOpen(false)}>
+                            Cancel
+                        </Button>
+                        <Button onClick={handleSubmit} disabled={!tempName.trim() || !tempEmail.trim()}>
+                            Submit
                         </Button>
                     </DialogFooter>
-                  </DialogContent>
-                </Dialog>
-              
+                </DialogContent>
+            </Dialog>
 
-    </>
-);
+
+        </>
+    );
 
 }
